requirements.txt
# Google APIs
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client
httpx>=0.26.0

# Database & storage
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Scheduling
apscheduler>=3.10.0

# Security
python-dotenv>=1.0.0
itsdangerous>=2.1.0

# MCP (Model Context Protocol) - custom implementation (mcp SDK requires Python 3.10+)
